Fix: When creating a new event, automatically update the user's joined Events on the sidebar.
